requirements.txt
dj-rest-auth
django-allauth==0.63.3
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0